+++ faddr/logging.py
@@ -6,11 +6,9 @@ from loguru import logger
 
 from faddr.settings import FaddrSettings
 
+
 settings = FaddrSettings()
 
 # Update logging level from settings
 logger.remove()
-if settings.debug:
-    logger.add(sys.stdout, level="DEBUG")
-else:
-    logger.add(sys.stdout, level="INFO")
+logger.add(sys.stdout, level=settings.log_level)