+++ tests/test_init.py
@@ -5,4 +5,4 @@ from faddr import __version__
 
 def test_version():
     """Test app version."""
-    assert __version__ == "0.4.0-alpha.1"
+    assert __version__ == "0.4.0-alpha.2"