+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Repo` class to unify config file handling
 - `RepoList` class to aggregate repos
 - `Config` dataclass to store config file data
+- REST API runner can be configured via environment variables
 
 ### Changed