+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2022-05-16
+
+### Fixed
+
+- Removed debug line that was printed even with `FADDR_DEBUG=false`
+
 ## [0.4.0] - 2022-05-15
 
 ### Added